# CUDA: faster dequantize kernels for Q4_0 and Q4_1