On my GPU (RTX-4080) prompt processing (PP) is faster for `Q4_K` than it is for `Q4_0` and `Q4_1`. This is unexpected as `Q4_K` de-quantization has more to do than `Q4_0`. So, I went ahead and implemented the same approach as is used for `Q4_K` also for `Q4_0`. I get the following results for `Q4_0` (`Q4_1` is very similar):

| model                          |       size |     params | backend    | test       |      t/s (Master)|        t/s (PR)  |  ratio |
| ------------------------------ | ---------: | ---------: | ---------- | ---------- | ---------------: | ---------------: | -----: |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       | pp 64      |   1503.23 ± 0.97 |   1603.60 ± 1.47 | 1.066  |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       | pp 128     |   2575.35 ± 0.78 |   2729.26 ± 1.86 | 1.060  |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       | pp 256     |   4236.93 ± 1.05 |   4489.80 ± 1.93 | 1.060  |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       | pp 512     |   5643.46 ± 3.20 |   5934.50 ± 1.79 | 1.052  |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       | pp 1024    |   5186.04 ± 4.66 |   5460.50 ± 5.18 | 1.053  |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       | pp 2048    |   4561.84 ± 3.12 |   4740.26 ± 3.55 | 1.039  |
| phi2 3B Q4_0                   |   1.49 GiB |     2.78 B | CUDA       | pp 64      |  3737.06 ± 73.80 | 4169.26 ± 138.48 | 1.116  |
| phi2 3B Q4_0                   |   1.49 GiB |     2.78 B | CUDA       | pp 128     |  6153.63 ± 13.18 |  6818.39 ± 13.93 | 1.108  |
| phi2 3B Q4_0                   |   1.49 GiB |     2.78 B | CUDA       | pp 256     |  8901.27 ± 16.68 |  9418.51 ± 13.08 | 1.058  |
| phi2 3B Q4_0                   |   1.49 GiB |     2.78 B | CUDA       | pp 512     | 10823.46 ± 10.70 |  11140.56 ± 9.31 | 1.029  |
| phi2 3B Q4_0                   |   1.49 GiB |     2.78 B | CUDA       | pp 1024    |   9499.67 ± 9.30 |  9775.45 ± 16.06 | 1.029  |
| phi2 3B Q4_0                   |   1.49 GiB |     2.78 B | CUDA       | pp 2048    |   7910.42 ± 9.57 |   8074.38 ± 8.51 | 1.021  |

If this change is considered acceptable, I can add similar kernels for `Q5_0` and `Q5_1`.

We see benefit is increasing with decreasing batch size, so it would be interesting to see results below 64 tokens, but I did not want to go and change the current small-batch approach (which uses the MMQ kernels).