@kalomaze 2024-01-14
The trend in the data seems to be that the smaller the batch, the less dequantization overhead there is. Makes me wonder if generation is impacted positively to some degree, but just wasn't measured