@JohannesGaessler 2024-01-15
>The trend in the data seems to be that the smaller the batch, the less dequantization overhead there is.

No it's not. Small batches have received a comparatively larger speedup therefore the conclusion is that small batches are spending a larger portion of the runtime on dequantization.