@ikawrakow 2024-01-15
> > The trend in the data seems to be that the smaller the batch, the less dequantization overhead there is.
> 
> No it's not. Small batches have received a comparatively larger speedup therefore the conclusion is that small batches are spending a larger portion of the runtime on dequantization.

Exactly.

On my GPU PP-512 for a 7B model is about 8000 t/s for `fp16`. This means that the GEMM takes about `512/8000` seconds per batch of 512, so about  64 ms (assuming all time is spent on matrix multiplications, and time taken by all other kernels is negligible). PP-512 for quantized models is about 6000 t/s, so about 85 ms per batch of 512. Memory bandwidth for my GPU is 768 GB/s, so best case scenario for writing the dequantized tensors to memory is `14/768` seconds, so about 18 ms. Add the 64 ms for GEMM, so 82 ms per batch of 512, or 6240 t/s, would be the very best achievable for PP-512 with a quantized model (implemented as dequantize fo `fp16` followed by `fp16` GEMM) on my GPU. Given this, the 6000 t/s I'm getting is not too bad. When batch size becomes smaller, we are still spending the 18+ ms to store the dequantized model in memory, but the GEMM is faster. Ignoring GPU utilization issues, GEMM is proportional to batch size, so for a batch of 64, GEMM is about 8 ms, so best possible total time on my GPU is `8+18 = 26` ms, so 2460 t/s. Instead, I get 1600 t/s, so we have significantly more inefficiency for smaller batches, and improving memory bandwidth utilization has a higher impact.