@JohannesGaessler 2024-01-15
>assuming all time is spent on matrix multiplications, and time taken by all other kernels is negligible

It's not quite as easy; last time I checked ~50% of the GPU time is spent on matrix multiplication for a batch size of 512 (with an RTX 3090). Among other things, there is significant overhead due to FP16<->FP32 conversions. You can check this quite easily yourself by just running `nsys profile <COMMAND>` and then opening the resulting file in NVIDIA NSight Systems. This is especially useful for optimizing small kernels where the impact on the overall runtime is small.

>When batch size becomes smaller, we are still spending the 18+ ms to store the dequantized model in memory, but the GEMM is faster.

I think this is the wrong way to think about it. For small batches the runtime of evaluating one batch decreases but the runtime *per token* increases because the overhead from dequantization is split across fewer tokens and it is more difficult to achieve good GPU compute utilization for small batches.

>Ignoring GPU utilization issues, GEMM is proportional to batch size

Especially for consumer GPUs this can **not** be assumed because NVIDIA likes to put awkward numbers of streaming multiprocessors onto them. If you for example take the 4080, it has 76 streaming multiprocessors. So for a typical computing grid size of 128 or 256 that already caps your GPU utilization to 84%. Assuming only powers of 2 you need a grid size of 512 to get a utilization of up to 96%.